Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
 int main(){
-    float a;
+    double a;
     
-    scanf("%f", &a);
-    printf("You entered: %.4f\n", a);
+    scanf("%lu", &a);
+    printf("You entered: %.4lu\n", a);
     return 0;
 }
